refactor: Spacer 다형성 컴포넌트 스타일 책임 분리 및 적용
