+++ src/pages/Login.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 import { ReactComponent as Kakao } from '../assets/svg/login/kakao.svg';
 import { ReactComponent as Naver } from '../assets/svg/login/naver.svg';
 import { ReactComponent as Logo } from '../assets/svg/login_main.svg';
+import Spacer from '../components/@common/Spacer';
 import Text from '../components/@common/Text/Text';
-import Spacer from '../components/Spacer';
 import UnStyleButton from '../components/UnStyleButton';
 
 const REQUEST_URL = `http://localhost:8080/oauth/kakao`;
@@ -17,7 +17,7 @@ const Login = () => {
   return (
     <Container>
       <Logo />
-      <Spacer top={40} />
+      <Spacer size={40} />
       <UnStyleButton onClick={onClickOauthKakao}>
         <ButtonWrapper>
           <LogoBox color="#FAE100">
@@ -29,7 +29,7 @@ const Login = () => {
         </ButtonWrapper>
       </UnStyleButton>
 
-      <Spacer top={10} />
+      <Spacer size={10} />
       <UnStyleButton>
         <ButtonWrapper>
           <LogoBox color="#03C75A">
